+++ main.go
@@ -6,10 +6,11 @@ import (
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/service/route53"
 	"github.com/urfave/cli"
+	"os"
 )
 
 var r53 *route53.Route53
-var version = "0.8.4"
+var version = "master"
 
 // Main entry point for cli53 application
 func Main(args []string) int {
@@ -25,6 +26,14 @@ func Main(args []string) int {
 		cli.StringFlag{
 			Name:  "profile",
 			Usage: "profile to use from credentials file",
+		},
+		cli.StringFlag{
+			Name:  "role-arn",
+			Usage: "AWS role ARN to assume",
+		},
+		cli.StringFlag{
+			Name:  "endpoint-url",
+			Usage: "override Route53 endpoint (hostname or fully qualified URI)",
 		},
 	}
 
@@ -44,8 +53,11 @@ func Main(args []string) int {
 					Usage: "output format: text, json, jl, table, csv",
 				},
 			),
-			Action: func(c *cli.Context) error {
-				r53 = getService(c)
+			Action: func(c *cli.Context) (err error) {
+				r53, err = getService(c)
+				if err != nil {
+					return err
+				}
 				if len(c.Args()) != 0 {
 					cli.ShowCommandHelp(c, "list")
 					return cli.NewExitError("No parameters expected", 1)
@@ -85,8 +97,11 @@ func Main(args []string) int {
 					Usage: "use the given delegation set",
 				},
 			),
-			Action: func(c *cli.Context) error {
-				r53 = getService(c)
+			Action: func(c *cli.Context) (err error) {
+				r53, err = getService(c)
+				if err != nil {
+					return err
+				}
 				if len(c.Args()) != 1 {
 					cli.ShowCommandHelp(c, "create")
 					return cli.NewExitError("Expected exactly 1 parameter", 1)
@@ -105,8 +120,11 @@ func Main(args []string) int {
 					Usage: "remove any existing records on the domain (otherwise deletion will fail)",
 				},
 			),
-			Action: func(c *cli.Context) error {
-				r53 = getService(c)
+			Action: func(c *cli.Context) (err error) {
+				r53, err = getService(c)
+				if err != nil {
+					return err
+				}
 				if len(c.Args()) != 1 {
 					cli.ShowCommandHelp(c, "delete")
 					return cli.NewExitError("Expected exactly 1 parameter", 1)
@@ -124,7 +142,7 @@ func Main(args []string) int {
 				cli.StringFlag{
 					Name:  "file",
 					Value: "",
-					Usage: "bind zone file (required)",
+					Usage: "bind zone filename, or - for stdin (required)",
 				},
 				cli.BoolFlag{
 					Name:  "wait",
@@ -143,8 +161,11 @@ func Main(args []string) int {
 					Usage: "perform a trial run with no changes made",
 				},
 			),
-			Action: func(c *cli.Context) error {
-				r53 = getService(c)
+			Action: func(c *cli.Context) (err error) {
+				r53, err = getService(c)
+				if err != nil {
+					return err
+				}
 				if len(c.Args()) != 1 {
 					cli.ShowCommandHelp(c, "import")
 					return cli.NewExitError("Expected exactly 1 parameter", 1)
@@ -162,6 +183,76 @@ func Main(args []string) int {
 			},
 		},
 		{
+			Name:      "instances",
+			Usage:     "dynamically update your dns with EC2 instance names",
+			ArgsUsage: "name|ID",
+			Flags: append(commonFlags,
+				cli.StringFlag{
+					Name:  "off",
+					Value: "",
+					Usage: "if provided, then records for stopped instances will be updated. This option gives the dns name the CNAME should revert to",
+				},
+				cli.StringSliceFlag{
+					Name:   "region",
+					EnvVar: "AWS_REGION",
+					Usage:  "a list of regions to check",
+				},
+				cli.BoolFlag{
+					Name:  "wait",
+					Usage: "wait for changes to become live",
+				},
+				cli.IntFlag{
+					Name:  "ttl, x",
+					Value: 60,
+					Usage: "resource record ttl",
+				},
+				cli.StringFlag{
+					Name:  "match",
+					Value: "",
+					Usage: "regular expression to select which Name tags to use",
+				},
+				cli.BoolFlag{
+					Name:  "internal, i",
+					Usage: "always use the internal hostname",
+				},
+				cli.BoolFlag{
+					Name:  "a-record, a",
+					Usage: "write an A record (IP) instead of CNAME",
+				},
+				cli.BoolFlag{
+					Name:  "dry-run, n",
+					Usage: "dry run - don't make any changes",
+				},
+			),
+			Action: func(c *cli.Context) (err error) {
+				config, err := getConfig(c)
+				if err != nil {
+					return err
+				}
+				r53, err = getService(c)
+				if err != nil {
+					return err
+				}
+				if len(c.Args()) != 1 {
+					cli.ShowCommandHelp(c, "instances")
+					return cli.NewExitError("Expected exactly 1 parameter", 1)
+				}
+				args := instancesArgs{
+					name:     c.Args().First(),
+					off:      c.String("off"),
+					regions:  c.StringSlice("region"),
+					wait:     c.Bool("wait"),
+					ttl:      c.Int("ttl"),
+					match:    c.String("match"),
+					internal: c.Bool("internal"),
+					aRecord:  c.Bool("a-record"),
+					dryRun:   c.Bool("dry-run"),
+				}
+				instances(args, config)
+				return nil
+			},
+		},
+		{
 			Name:      "export",
 			Usage:     "export a bind zone file (to stdout)",
 			ArgsUsage: "name|ID",
@@ -170,14 +261,31 @@ func Main(args []string) int {
 					Name:  "full, f",
 					Usage: "export prefixes as full names",
 				},
+				cli.StringFlag{
+					Name:  "output",
+					Usage: "Write to an output file instead of STDOUT",
+				},
 			),
-			Action: func(c *cli.Context) error {
-				r53 = getService(c)
+			Action: func(c *cli.Context) (err error) {
+				r53, err = getService(c)
+				if err != nil {
+					return err
+				}
 				if len(c.Args()) != 1 {
 					cli.ShowCommandHelp(c, "export")
 					return cli.NewExitError("Expected exactly 1 parameter", 1)
 				}
-				exportBind(c.Args().First(), c.Bool("full"))
+
+				outputFileName := c.String("output")
+				writer := os.Stdout
+				if len(outputFileName) > 0 {
+					writer, err = os.Create(outputFileName)
+					defer writer.Close()
+					if err != nil {
+						return err
+					}
+				}
+				exportBind(c.Args().First(), c.Bool("full"), writer)
 				return nil
 			},
 		},
@@ -231,9 +339,16 @@ func Main(args []string) int {
 					Name:  "subdivision-code",
 					Usage: "subdivision code for geolocation routing",
 				},
+				cli.BoolFlag{
+					Name:  "multivalue",
+					Usage: "use multivalue answer routing",
+				},
 			),
-			Action: func(c *cli.Context) error {
-				r53 = getService(c)
+			Action: func(c *cli.Context) (err error) {
+				r53, err = getService(c)
+				if err != nil {
+					return err
+				}
 				if len(c.Args()) < 2 {
 					cli.ShowCommandHelp(c, "rrcreate")
 					return cli.NewExitError("Expected at least 2 parameters", 1)
@@ -256,6 +371,7 @@ func Main(args []string) int {
 					countryCode:     c.String("country-code"),
 					continentCode:   c.String("continent-code"),
 					subdivisionCode: c.String("subdivision-code"),
+					multivalue:      c.Bool("multivalue"),
 				}
 				if args.validate() {
 					createRecords(args)
@@ -280,8 +396,11 @@ func Main(args []string) int {
 					Usage: "record set identifier to delete",
 				},
 			),
-			Action: func(c *cli.Context) error {
-				r53 = getService(c)
+			Action: func(c *cli.Context) (err error) {
+				r53, err = getService(c)
+				if err != nil {
+					return err
+				}
 				if len(c.Args()) != 3 {
 					cli.ShowCommandHelp(c, "rrdelete")
 					return cli.NewExitError("Expected exactly 3 parameters", 1)
@@ -304,8 +423,11 @@ func Main(args []string) int {
 					Usage: "wait for changes to become live",
 				},
 			),
-			Action: func(c *cli.Context) error {
-				r53 = getService(c)
+			Action: func(c *cli.Context) (err error) {
+				r53, err = getService(c)
+				if err != nil {
+					return err
+				}
 				if len(c.Args()) != 1 {
 					cli.ShowCommandHelp(c, "rrpurge")
 					return cli.NewExitError("Expected exactly 1 parameter", 1)
@@ -321,8 +443,11 @@ func Main(args []string) int {
 			Name:  "dslist",
 			Usage: "list reusable delegation sets",
 			Flags: commonFlags,
-			Action: func(c *cli.Context) error {
-				r53 = getService(c)
+			Action: func(c *cli.Context) (err error) {
+				r53, err = getService(c)
+				if err != nil {
+					return err
+				}
 				listReusableDelegationSets()
 				return nil
 			},
@@ -337,8 +462,11 @@ func Main(args []string) int {
 					Usage: "convert the given zone delegation set (optional)",
 				},
 			),
-			Action: func(c *cli.Context) error {
-				r53 = getService(c)
+			Action: func(c *cli.Context) (err error) {
+				r53, err = getService(c)
+				if err != nil {
+					return err
+				}
 				createReusableDelegationSet(c.String("zone-id"))
 				return nil
 			},
@@ -348,8 +476,11 @@ func Main(args []string) int {
 			Usage:     "delete a reusable delegation set",
 			ArgsUsage: "id",
 			Flags:     commonFlags,
-			Action: func(c *cli.Context) error {
-				r53 = getService(c)
+			Action: func(c *cli.Context) (err error) {
+				r53, err = getService(c)
+				if err != nil {
+					return err
+				}
 				if len(c.Args()) != 1 {
 					cli.ShowCommandHelp(c, "dsdelete")
 					return cli.NewExitError("Expected exactly 1 parameter", 1)